remove secrets class
